zestimator-backend/requirements.txt
flask
requests
python-dotenv
playwright
flask-sqlalchemy
flask-migrate
flask-socketio
flask-cors
psycopg2-binary
